fix merge bugs
