+++ collector/export_test.go
@@ -1,0 +1,3 @@
+package collector
+
+var ProcessSingleMessage = processSingleMessage
